pyproject.toml
[build-system]
requires = ["setuptools>=65"]
build-backend = "setuptools.build_meta"

[project]
name = "lineprophet-machine-learning-for-spectral-attributes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic==1.14.0",
    "anyio==4.7.0",
    "argon2-cffi==23.1.0",
    "argon2-cffi-bindings==21.2.0",
    "arrow==1.3.0",
    "asteval==1.0.5",
    "asttokens==3.0.0",
    "async-lru==2.0.4",
    "attrs==24.3.0",
    "babel==2.16.0",
    "beautifulsoup4==4.12.3",
    "bleach==6.2.0",
    "certifi==2024.12.14",
    "cffi==1.17.1",
    "charset-normalizer==3.4.1",
    "colorlog==6.9.0",
    "comm==0.2.2",
    "contourpy==1.3.1",
    "cycler==0.12.1",
    "debugpy==1.8.11",
    "decorator==5.1.1",
    "defusedxml==0.7.1",
    "dill==0.3.9",
    "exceptiongroup==1.2.2",
    "executing==2.1.0",
    "fastjsonschema==2.21.1",
    "filelock==3.17.0",
    "fonttools==4.55.3",
    "fqdn==1.5.1",
    "fsspec==2024.12.0",
    "greenlet==3.1.1",
    "h11==0.14.0",
    "hitran-api==1.2.2.3",
    "httpcore==1.0.7",
    "httpx==0.28.1",
    "idna==3.10",
    "ipykernel==6.29.5",
    "ipython==8.31.0",
    "ipywidgets==8.1.5",
    "isoduration==20.11.0",
    "jedi==0.19.2",
    "jinja2==3.1.5",
    "json5==0.10.0",
    "jsonpointer==3.0.0",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2024.10.1",
    "jupyter==1.1.1",
    "jupyter-client==8.6.3",
    "jupyter-console==6.6.3",
    "jupyter-core==5.7.2",
    "jupyter-events==0.11.0",
    "jupyter-lsp==2.2.5",
    "jupyter-server==2.15.0",
    "jupyter-server-terminals==0.5.3",
    "jupyterlab==4.3.4",
    "jupyterlab-pygments==0.3.0",
    "jupyterlab-server==2.27.3",
    "jupyterlab-widgets==3.0.13",
    "kiwisolver==1.4.8",
    "lightgbm>=4.6.0",
    "llvmlite==0.43.0",
    "lmfit==1.3.2",
    "mako==1.3.8",
    "markupsafe==3.0.2",
    "matplotlib==3.10.0",
    "matplotlib-inline==0.1.7",
    "mistune==3.0.2",
    "mpmath==1.3.0",
    "nbclient==0.10.2",
    "nbconvert==7.16.4",
    "nbformat==5.10.4",
    "nest-asyncio==1.6.0",
    "networkx==3.4.2",
    "notebook==7.3.2",
    "notebook-shim==0.2.4",
    "numba==0.60.0",
    "numpy==2.0.2",
    "nvidia-cublas-cu12==12.4.5.8",
    "nvidia-cuda-cupti-cu12==12.4.127",
    "nvidia-cuda-nvrtc-cu12==12.4.127",
    "nvidia-cuda-runtime-cu12==12.4.127",
    "nvidia-cudnn-cu12==9.1.0.70",
    "nvidia-cufft-cu12==11.2.1.3",
    "nvidia-curand-cu12==10.3.5.147",
    "nvidia-cusolver-cu12==11.6.1.9",
    "nvidia-cusparse-cu12==12.3.1.170",
    "nvidia-nccl-cu12==2.21.5",
    "nvidia-nvjitlink-cu12==12.4.127",
    "nvidia-nvtx-cu12==12.4.127",
    "optuna==4.1.0",
    "overrides==7.7.0",
    "packaging==24.2",
    "pandas==2.2.3",
    "pandocfilters==1.5.1",
    "parso==0.8.4",
    "pexpect==4.9.0",
    "pillow==11.0.0",
    "platformdirs==4.3.6",
    "plotly==5.24.1",
    "prometheus-client==0.21.1",
    "prompt-toolkit==3.0.48",
    "psutil==6.1.1",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pycparser==2.22",
    "pygments==2.18.0",
    "pyparsing==3.2.0",
    "pyqt5-qt5==5.15.16",
    "pyqt5-sip==12.16.1",
    "pyside2==5.15.2.1",
    "python-dateutil==2.9.0.post0",
    "python-json-logger==3.2.1",
    "pytz==2024.2",
    "pyyaml==6.0.2",
    "pyzmq==26.2.0",
    "referencing==0.35.1",
    "requests==2.32.3",
    "rfc3339-validator==0.1.4",
    "rfc3986-validator==0.1.1",
    "rpds-py==0.22.3",
    "scikit-learn>=1.6.1",
    "scipy==1.14.1",
    "seaborn==0.13.2",
    "send2trash==1.8.3",
    "shiboken2==5.15.2.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.6",
    "sqlalchemy==2.0.36",
    "stack-data==0.6.3",
    "sympy==1.13.1",
    "tabulate==0.9.0",
    "tenacity==9.0.0",
    "tensorboard>=2.18.0",
    "terminado==0.18.1",
    "tinycss2==1.4.0",
    "tk==0.1.0",
    "tomli==2.2.1",
    "torch==2.5.1",
    "torchinfo==1.8.0",
    "torchvision==0.20.1",
    "tornado==6.4.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "triton==3.1.0",
    "types-python-dateutil==2.9.0.20241206",
    "typing-extensions==4.12.2",
    "tzdata==2024.2",
    "uncertainties==3.2.2",
    "uri-template==1.3.0",
    "urllib3==2.3.0",
    "wcwidth==0.2.13",
    "webcolors==24.11.1",
    "webencodings==0.5.1",
    "websocket-client==1.8.0",
    "widgetsnbextension==4.0.13",
    "xgboost>=2.1.4",
]

[tool.setuptools.packages]
find = { where = ["src"] }  # Procura pacotes dentro de src/
